TACO-35 address pr feedback - remove duplicate functionality in OkHttpUnsafe
