+++ xio-core/pom.xml
@@ -242,6 +242,13 @@
     </dependency>
 
     <dependency>
+      <groupId>com.xjeffrose</groupId>
+      <artifactId>xio-test</artifactId>
+      <version>0.13.4-SNAPSHOT</version>
+      <scope>test</scope>
+    </dependency>
+
+    <dependency>
       <groupId>org.mockito</groupId>
       <artifactId>mockito-all</artifactId>
       <version>1.10.19</version>
