+++ xio-test/src/main/java/com/xjeffrose/xio/test/OkHttpUnsafe.java
@@ -7,6 +7,8 @@ import java.security.PrivateKey;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
 import javax.net.ssl.HostnameVerifier;
 import javax.net.ssl.KeyManager;
 import javax.net.ssl.KeyManagerFactory;
@@ -38,26 +40,14 @@ public class OkHttpUnsafe {
     return keyManagerFactory.getKeyManagers();
   }
 
-  public static KeyManager[] getEmptyKeyManager() throws Exception {
-    KeyStore keystore = KeyStore.getInstance(KeyStore.getDefaultType());
-    keystore.load(null, "".toCharArray());
-    KeyManagerFactory keyManagerFactory =
-        KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
-    keyManagerFactory.init(keystore, "".toCharArray());
-    return keyManagerFactory.getKeyManagers();
-  }
-
   public static X509TrustManager unsafeTrustManager() {
     return new X509TrustManager() {
-      @Override
       public void checkClientTrusted(java.security.cert.X509Certificate[] chain, String authType)
           throws CertificateException {}
 
-      @Override
       public void checkServerTrusted(java.security.cert.X509Certificate[] chain, String authType)
           throws CertificateException {}
 
-      @Override
       public java.security.cert.X509Certificate[] getAcceptedIssuers() {
         return new X509Certificate[0];
       }
@@ -84,24 +74,25 @@ public class OkHttpUnsafe {
     return sslSocketFactory;
   }
 
-  public static OkHttpClient getUnsafeClient() throws Exception {
+  public static OkHttpClient getUnsafeClient(Protocol... protocols) throws Exception {
     X509TrustManager trustManager = unsafeTrustManager();
     final SSLSocketFactory sslSocketFactory = getUnsafeSSLSocketFactory(null, trustManager);
-
-    OkHttpClient okHttpClient =
-        new OkHttpClient.Builder()
-            .sslSocketFactory(sslSocketFactory, trustManager)
-            .hostnameVerifier(
-                new HostnameVerifier() {
-                  @Override
-                  public boolean verify(String hostname, SSLSession session) {
-                    return true;
-                  }
-                })
-            .protocols(Arrays.asList(Protocol.HTTP_1_1))
-            .build();
-
-    return okHttpClient;
+    final List<Protocol> protocolList;
+    if (protocols.length == 0) {
+      protocolList = Collections.singletonList(Protocol.HTTP_1_1);
+    } else {
+      protocolList = Arrays.asList(protocols);
+    }
+    return new OkHttpClient.Builder()
+        .sslSocketFactory(sslSocketFactory, trustManager)
+        .hostnameVerifier(
+            new HostnameVerifier() {
+              public boolean verify(String hostname, SSLSession session) {
+                return true;
+              }
+            })
+        .protocols(protocolList)
+        .build();
   }
 
   public static MockWebServer getSslMockWebServer(KeyManager[] keyManagers) throws Exception {
@@ -109,10 +100,5 @@ public class OkHttpUnsafe {
     server.useHttps(
         OkHttpUnsafe.getUnsafeSSLSocketFactory(keyManagers, unsafeTrustManager()), false);
     return server;
-  }
-
-  public static MockWebServer getSslMockWebServer(
-      PrivateKey privateKey, X509Certificate[] certificateAndChain) throws Exception {
-    return getSslMockWebServer(getKeyManagers(privateKey, certificateAndChain));
   }
 }
